move QuizAttribute to manager
